requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.117
botocore==1.34.117
celery==5.4.0
celery-progress==0.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.1
Django==4.2.13
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
importlib_resources==6.4.0
incremental==22.10.0
gunicorn==22.0.0
isodate==0.6.1
jmespath==1.0.1
jwt==1.3.1
kombu==5.3.7
logzero==1.7.0
numpy==1.24.4
openpyxl==3.1.3
pandas==2.0.3
pendulum==3.0.0
prompt_toolkit==3.0.46
py4j==0.10.9.7
pycparser==2.21
PyJWT==2.8.0
pyotp==2.8.0
pyparsing==3.1.0
pyspark==3.5.1
pytz==2023.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
redis==5.0.4
requests==2.31.0
s3transfer==0.10.1
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
sqlparse==0.5.0
time-machine==2.14.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2024.1
urllib3==1.25.4
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.6.0
zipp==3.19.2
zope.interface==6.0